Split rolling checksum tests into modules
