+++ crates/checksums/src/rolling/tests/checksum/basic.rs
@@ -1,8 +1,5 @@
 use super::super::*;
-use super::{chunked_sequences, random_data_and_window, roll_many_sequences};
 
-use proptest::prelude::*;
-use std::collections::VecDeque;
 use std::io::{self, Cursor, IoSlice, Read};
 
 #[test]
@@ -346,261 +343,4 @@ fn update_vectored_coalesces_small_slices() {
 
     assert_eq!(vectored.digest(), sequential.digest());
     assert_eq!(vectored.value(), sequential.value());
-}
-
-proptest! {
-    #[test]
-    fn rolling_update_matches_single_pass(chunks in chunked_sequences()) {
-        let mut incremental = RollingChecksum::new();
-        let mut concatenated = Vec::new();
-
-        for chunk in &chunks {
-            incremental.update(chunk);
-            concatenated.extend_from_slice(chunk);
-        }
-
-        let mut single_pass = RollingChecksum::new();
-        single_pass.update(&concatenated);
-
-        prop_assert_eq!(incremental.digest(), single_pass.digest());
-        prop_assert_eq!(incremental.value(), single_pass.value());
-    }
-
-    #[test]
-    fn rolling_matches_reference_for_random_windows((data, window) in random_data_and_window()) {
-        let mut rolling = RollingChecksum::new();
-        rolling.update(&data[..window]);
-
-        let mut reference = RollingChecksum::new();
-        reference.update(&data[..window]);
-
-        prop_assert_eq!(rolling.digest(), reference.digest());
-        prop_assert_eq!(rolling.value(), reference.value());
-
-        if data.len() > window {
-            for start in 1..=data.len() - window {
-                let outgoing = data[start - 1];
-                let incoming = data[start + window - 1];
-                rolling
-                    .roll(outgoing, incoming)
-                    .expect("rolling update must succeed");
-
-                let mut recomputed = RollingChecksum::new();
-                recomputed.update(&data[start..start + window]);
-
-                prop_assert_eq!(rolling.digest(), recomputed.digest());
-                prop_assert_eq!(rolling.value(), recomputed.value());
-            }
-        }
-    }
-
-    #[test]
-    fn vectored_update_matches_chunked_input(chunks in chunked_sequences()) {
-        let mut sequential = RollingChecksum::new();
-        for chunk in &chunks {
-            sequential.update(chunk);
-        }
-
-        let slices: Vec<IoSlice<'_>> =
-            chunks.iter().map(|chunk| IoSlice::new(chunk.as_slice())).collect();
-
-        let mut vectored = RollingChecksum::new();
-        vectored.update_vectored(&slices);
-
-        prop_assert_eq!(vectored.digest(), sequential.digest());
-        prop_assert_eq!(vectored.value(), sequential.value());
-    }
-
-    #[test]
-    fn roll_many_matches_single_rolls_for_random_sequences(
-        (seed, pairs) in roll_many_sequences(),
-    ) {
-        let mut optimized = RollingChecksum::new();
-        optimized.update(&seed);
-
-        let mut reference = optimized.clone();
-
-        let (outgoing, incoming): (Vec<u8>, Vec<u8>) = pairs.into_iter().unzip();
-        optimized
-            .roll_many(&outgoing, &incoming)
-            .expect("multi-byte roll succeeds");
-
-        for (&out, &inn) in outgoing.iter().zip(incoming.iter()) {
-            reference
-                .roll(out, inn)
-                .expect("single-byte roll succeeds");
-        }
-
-        prop_assert_eq!(optimized.digest(), reference.digest());
-        prop_assert_eq!(optimized.value(), reference.value());
-    }
-
-
-    #[test]
-    fn from_digest_round_trips(data in prop::collection::vec(any::<u8>(), 0..=256)) {
-        let mut checksum = RollingChecksum::new();
-        checksum.update(&data);
-
-        let digest = checksum.digest();
-        let restored = RollingChecksum::from_digest(digest);
-
-        prop_assert_eq!(restored.digest(), digest);
-        prop_assert_eq!(restored.value(), checksum.value());
-        prop_assert_eq!(restored.len(), checksum.len());
-    }
-}
-
-#[test]
-fn roll_many_matches_single_rolls_for_long_sequences() {
-    let seed: Vec<u8> = (0..128)
-        .map(|value| {
-            let byte = value as u8;
-            byte.wrapping_mul(13).wrapping_add(5)
-        })
-        .collect();
-
-    let mut batched = RollingChecksum::new();
-    batched.update(&seed);
-
-    let mut reference = batched.clone();
-
-    let mut window = VecDeque::from(seed.clone());
-    let mut outgoing = Vec::with_capacity(4096);
-    let mut incoming = Vec::with_capacity(4096);
-
-    for step in 0..4096 {
-        let leaving = window
-            .pop_front()
-            .expect("rolling checksum window must contain data");
-        let step_byte = step as u8;
-        let entering = step_byte
-            .wrapping_mul(17)
-            .wrapping_add(23)
-            .wrapping_add(step_byte >> 3);
-
-        outgoing.push(leaving);
-        incoming.push(entering);
-        window.push_back(entering);
-    }
-
-    batched
-        .roll_many(&outgoing, &incoming)
-        .expect("batched roll succeeds");
-
-    for (&out, &inn) in outgoing.iter().zip(incoming.iter()) {
-        reference.roll(out, inn).expect("sequential roll succeeds");
-    }
-
-    assert_eq!(batched.digest(), reference.digest());
-    assert_eq!(batched.value(), reference.value());
-}
-
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-#[test]
-fn sse2_accumulate_matches_scalar_reference() {
-    if !std::arch::is_x86_feature_detected!("sse2") {
-        return;
-    }
-
-    use crate::rolling::checksum::accumulate_chunk_scalar_for_tests;
-    use crate::rolling::checksum::x86::accumulate_chunk_sse2_for_tests;
-
-    let sizes = [1usize, 15, 16, 17, 63, 64, 65, 128, 511, 4096];
-    let seeds = [
-        (0u32, 0u32, 0usize),
-        (0x1234u32, 0x5678u32, 7usize),
-        (0x0fffu32, 0x7fffu32, 1024usize),
-        (0xffffu32, 0xffffu32, usize::MAX - 32),
-    ];
-
-    for &(seed_s1, seed_s2, seed_len) in &seeds {
-        for &size in &sizes {
-            let mut data = vec![0u8; size];
-            for (idx, byte) in data.iter_mut().enumerate() {
-                *byte = (idx as u8)
-                    .wrapping_mul(31)
-                    .wrapping_add((size as u8).wrapping_mul(3));
-            }
-
-            let scalar = accumulate_chunk_scalar_for_tests(seed_s1, seed_s2, seed_len, &data);
-            let simd = accumulate_chunk_sse2_for_tests(seed_s1, seed_s2, seed_len, &data);
-
-            assert_eq!(
-                scalar, simd,
-                "SSE2 mismatch for size {size} with seeds {seed_s1:#x}/{seed_s2:#x}/{seed_len}",
-            );
-        }
-    }
-}
-
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-#[test]
-fn avx2_accumulate_matches_scalar_reference() {
-    if !std::arch::is_x86_feature_detected!("avx2") {
-        return;
-    }
-
-    use crate::rolling::checksum::accumulate_chunk_scalar_for_tests;
-    use crate::rolling::checksum::x86::accumulate_chunk_avx2_for_tests;
-
-    let sizes = [32usize, 33, 47, 64, 95, 128, 1024, 4096];
-    let seeds = [
-        (0u32, 0u32, 0usize),
-        (0x1234u32, 0x5678u32, 7usize),
-        (0x0fffu32, 0x7fffu32, 1024usize),
-        (0xffffu32, 0xffffu32, usize::MAX - 64),
-    ];
-
-    for &(seed_s1, seed_s2, seed_len) in &seeds {
-        for &size in &sizes {
-            let mut data = vec![0u8; size];
-            for (idx, byte) in data.iter_mut().enumerate() {
-                *byte = (idx as u8)
-                    .wrapping_mul(17)
-                    .wrapping_add((size as u8).wrapping_mul(5));
-            }
-
-            let scalar = accumulate_chunk_scalar_for_tests(seed_s1, seed_s2, seed_len, &data);
-            let simd = accumulate_chunk_avx2_for_tests(seed_s1, seed_s2, seed_len, &data);
-
-            assert_eq!(
-                scalar, simd,
-                "AVX2 mismatch for size {size} with seeds {seed_s1:#x}/{seed_s2:#x}/{seed_len}",
-            );
-        }
-    }
-}
-
-#[cfg(target_arch = "aarch64")]
-#[test]
-fn neon_accumulate_matches_scalar_reference() {
-    use crate::rolling::checksum::accumulate_chunk_scalar_for_tests;
-    use crate::rolling::checksum::neon::accumulate_chunk_neon_for_tests;
-
-    let sizes = [1usize, 15, 16, 17, 63, 64, 65, 128, 511, 4096];
-    let seeds = [
-        (0u32, 0u32, 0usize),
-        (0x1234u32, 0x5678u32, 7usize),
-        (0x0fffu32, 0x7fffu32, 1024usize),
-        (0xffffu32, 0xffffu32, usize::MAX - 32),
-    ];
-
-    for &(seed_s1, seed_s2, seed_len) in &seeds {
-        for &size in &sizes {
-            let mut data = vec![0u8; size];
-            for (idx, byte) in data.iter_mut().enumerate() {
-                *byte = (idx as u8)
-                    .wrapping_mul(29)
-                    .wrapping_add((size as u8).wrapping_mul(5));
-            }
-
-            let scalar = accumulate_chunk_scalar_for_tests(seed_s1, seed_s2, seed_len, &data);
-            let simd = accumulate_chunk_neon_for_tests(seed_s1, seed_s2, seed_len, &data);
-
-            assert_eq!(
-                scalar, simd,
-                "NEON mismatch for size {size} with seeds {seed_s1:#x}/{seed_s2:#x}/{seed_len}",
-            );
-        }
-    }
 }
