+++ crates/checksums/src/rolling/tests/checksum/mod.rs
@@ -1,0 +1,5 @@
+#![deny(unsafe_code)]
+
+mod basic;
+mod properties;
+mod simd;
